[OC-870] Refactoring cardWriteSerice class in cardPublication service
