+++ services/core/cards-publication/src/main/java/org/lfenergy/operatorfabric/cards/publication/CardPublicationApplication.java
@@ -16,14 +16,12 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.ImportResource;
-import org.springframework.data.mongodb.repository.config.EnableReactiveMongoRepositories;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 
 @SpringBootApplication
 @Slf4j
 @RefreshScope
 @EnableOperatorFabricMongo
-@EnableReactiveMongoRepositories
 @ImportResource("classpath:/amqp.xml")
 public class CardPublicationApplication {
