+++ services/core/cards-publication/src/test/java/org/lfenergy/operatorfabric/cards/publication/repositories/CardRepositoryForTest.java
@@ -8,17 +8,17 @@
 
 package org.lfenergy.operatorfabric.cards.publication.repositories;
 
-import org.lfenergy.operatorfabric.cards.publication.model.ArchivedCardPublicationData;
+import org.lfenergy.operatorfabric.cards.publication.model.CardPublicationData;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import org.springframework.stereotype.Repository;
-import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 /**
- * <p>Auto generated spring mongo reactive repository to access Archived Card collection</p>
+ * <p>Auto generated spring mongo reactive repository to access Card collection</p>
  *
  */
 @Repository
-public interface ArchivedCardRepository extends ReactiveMongoRepository<ArchivedCardPublicationData,String> {
+public interface CardRepositoryForTest extends ReactiveMongoRepository<CardPublicationData,String> {
 
-    public Flux<ArchivedCardPublicationData> findByProcessId(String processId);
+    public Mono<CardPublicationData> findByProcessId(String processId);
 }
