+++ src/docs/asciidoc/docs/release_notes.adoc
@@ -31,3 +31,4 @@ Group{
 Please do take a look at the Users API documentation (groups section) :
 https://opfab.github.io/documentation/current/api/users/#/groups
 * [OC-858] Refactoring UI Timeline code for circle content and position computing
+* [OC-870] Refactoring cardWriteSerice class in cardPublication service
